backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for the Todo application"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "sqlmodel>=0.0.16",
    "uvicorn>=0.24.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",  # For PostgreSQL async driver
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.7",  # For sync database operations
    "python-jose>=3.3.0",  # For JWT handling
    "passlib>=1.7.4",  # For password hashing
    "argon2-cffi>=21.3.0",  # For password hashing algorithm
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.3.0",
    "openai>=2.15.0",
    "alembic>=1.18.1",
    "requests>=2.31.0",
    "aiosqlite>=0.19.0",
]
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]
test = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--strict-markers",
    "--strict-config",
]
markers = [
    "asyncio: marks tests as asyncio",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/conftest.py"]
